# f1api/openapi/api.yml
openapi: "3.0.0"
info:
  title: F1 API
  version: "1.0.0"
  description: |
    ## Paging and Sorting

    The following query paramters can be used to control the paging and sorting of collections.

    ### `pageSize`

    Defines the number of elements per page.

    **Default value:** 25

    **Notice:** maximum number is 100

    ### `pageNumber`

    Specifies the page number that should be returned.

    **Notice:** first page has the number 0

    **Default value:** 0

    ### `sort`

    Specifies by which properties to sort. <br>
    By default the sort order is ascending. This can be changed by adding `asc` for ascending or `desc` for
    descending separated by a comma after the sort property.

    **Hint:** To sort by more than one property, simply add a new sort parameter to the query.

    ### Example

    GET `/drivers?pageSize=20&pageNumber=0&sort=firstName,asc`

    will return the first page with a maximum of 20 drivers sorted by first name in ascending order.
...
paths:
...
  /races:
    get:
      summary: Get races / Get races by raceName / Get race by date
      tags:
        - "Races"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceName"
        - $ref: "./components/parameters.yml#/date"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Races"
        400:
          $ref: "./components/responses.yml#/400"
        500:
          $ref: "./components/responses.yml#/500"
  /races/{raceId}:
    get:
      summary: Get race
      tags:
        - "Races"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Race"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update race
      tags:
        - "Races"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Race"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Race"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
  /races/{raceId}/qualifying:
    post:
      x-internal: true
      summary: Create qualifying for a driver of a race
      description: |
        **Notice:** The driver must be in the team of the season to which the race belongs. <br>
        If the driver is not in the team, he can be added to it by patching the associated `teamOfSeason` entity. <br>
        **Notice:** Creates new `driverOfRace` entity for the given `raceId`, `driverId` and `teamId`.
      tags:
        - "Qualifyings"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Qualifying"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Qualifying"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get qualifying of a race / Get qualifying of a driver of a race by driverId
      tags:
        - "Qualifyings"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQuery"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Qualifyings"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update qualifying for a driver of a race
      tags:
        - "Qualifyings"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQueryRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/QualifyingPatch"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Qualifying"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    # Delete example for VOs
    #delete:
    #  summary: Delete qualifying for a driver of a race
    #  tags:
    #    - "Qualifyings"
    #    - "RaceController"
    #  parameters:
    #    - $ref: "./components/parameters.yml#/raceId"
    #    - $ref: "./components/parameters.yml#/driverIdQueryRequired"
    #  responses:
    #    204:
    #      $ref: "./components/responses.yml#/204"
    #    400:
    #      $ref: "./components/responses.yml#/400"
    #    404:
    #      $ref: "./components/responses.yml#/404"
    #    422:
    #      $ref: "./components/responses.yml#/422"
    #    409:
    #      $ref: "./components/responses.yml#/409"
    #    500:
    #      $ref: "./components/responses.yml#/500"#
  /races/{raceId}/result:
    post:
      x-internal: true
      summary: Create race result for a driver of a race
      description: |
        **Notice:** The driver must have a qualifying entry for the race.
      tags:
        - "Results"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/Result"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Result"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get race result of a race / Get race result of a driver of a race by driverId
      tags:
        - "Results"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQuery"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Results"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update race result for a driver of a race
      tags:
        - "Results"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQueryRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/ResultPatch"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/Result"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
  /races/{raceId}/fastestLaps:
    post:
      x-internal: true
      summary: Create fastestLap for a driver of a race
      description: |
        **Notice:** The driver must have a race result for the race.
      tags:
        - "FastestLaps"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/FastestLap"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/FastestLap"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get fastestLaps of a race / Get fastestLap of a driver of a race by driverId
      tags:
        - "FastestLaps"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQuery"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/FastestLaps"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update fastestLap for a driver of a race
      tags:
        - "FastestLaps"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQueryRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/FastestLapPatch"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/FastestLap"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
  /races/{raceId}/topSpeeds:
    post:
      x-internal: true
      summary: Create topSpeed for a driver of a race
      description: |
        **Notice:** The driver must have a race result for the race.
      tags:
        - "TopSpeeds"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/TopSpeed"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/TopSpeed"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get topSpeeds of a race / Get topSpeed of a driver of a race by driverId
      tags:
        - "TopSpeeds"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQuery"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/TopSpeeds"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update topSpeed for a driver of a race
      tags:
        - "TopSpeeds"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQueryRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/TopSpeedPatch"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/TopSpeed"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
  /races/{raceId}/leadingLaps:
    post:
      x-internal: true
      summary: Create leadingLaps for a driver of a race
      description: |
        **Notice:** The driver must have a race result for the race.
      tags:
        - "LeadingLaps"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/LeadingLaps"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/LeadingLaps"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get leadingLaps of a race / Get leadingLaps of a driver of a race by driverId
      tags:
        - "LeadingLaps"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQuery"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/LeadingLapsList"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update leadingLaps for a driver of a race
      tags:
        - "LeadingLaps"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
        - $ref: "./components/parameters.yml#/driverIdQueryRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/LeadingLapsPatch"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/LeadingLaps"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
  /races/{raceId}/fastestPitStop:
    post:
      x-internal: true
      summary: Create fastestPitStop of a race
      description: |
        **Notice:** There can only be one fastestPitStop of the race. <br>
        **Notice:** The driver must have a race result for the race.
      tags:
        - "FastestPitStops"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/FastestPitStop"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/FastestPitStop"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        409:
          $ref: "./components/responses.yml#/409"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"
    get:
      summary: Get fastestPitStop of a race
      tags:
        - "FastestPitStops"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/FastestPitStop"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        500:
          $ref: "./components/responses.yml#/500"
    patch:
      x-internal: true
      summary: Update fastestPitStop of a race
      tags:
        - "FastestPitStops"
        - "RaceController"
      parameters:
        - $ref: "./components/parameters.yml#/raceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./components/schemas.yml#/FastestPitStop"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yml#/FastestPitStop"
        400:
          $ref: "./components/responses.yml#/400"
        404:
          $ref: "./components/responses.yml#/404"
        422:
          $ref: "./components/responses.yml#/422"
        500:
          $ref: "./components/responses.yml#/500"

  # TeamController
...